+++ src/integrations/automations/WebRequest.php
@@ -54,6 +54,14 @@ class WebRequest extends Automation
 
         parent::__construct($config);
     }
+    
+    public function setAttributes($values, $safeOnly = true): void
+    {
+        // Handle legacy webhook from the form settings, which won't be via `__construct`
+        $values['url'] = ArrayHelper::remove($values, 'webhook');
+
+        parent::setAttributes($values, $safeOnly);
+    }
 
     public function getDescription(): string
     {